Use web3j SignatureData as signature internal representation
