+++ sdk/src/main/java/io/singularitynet/sdk/ethereum/PrivateKeyIdentity.java
@@ -1,12 +1,10 @@
 package io.singularitynet.sdk.ethereum;
 
 import java.math.BigInteger;
-import java.io.ByteArrayOutputStream;
 import org.web3j.crypto.ECKeyPair;
 import org.web3j.crypto.Sign;
 import org.web3j.crypto.Keys;
 import org.web3j.crypto.Hash;
-import static io.singularitynet.sdk.common.Preconditions.checkState;
 
 import io.singularitynet.sdk.common.Utils;
 
@@ -29,22 +27,12 @@ public class PrivateKeyIdentity implements Signer {
     @Override
     public Signature sign(byte[] message) {
         Sign.SignatureData signature = Sign.signPrefixedMessage(Hash.sha3(message), key);
-        return new Signature(signatureToBytes(signature));
+        return new Signature(signature);
     }
 
     @Override
     public Address getAddress() {
         return new Address(Keys.getAddress(key.getPublicKey()));
-    }
-
-    private static byte[] signatureToBytes(Sign.SignatureData signature) {
-        return Utils.wrapExceptions(() -> {
-            ByteArrayOutputStream bytes = new ByteArrayOutputStream();
-            bytes.write(signature.getR());
-            bytes.write(signature.getS());
-            bytes.write(signature.getV());
-            return bytes.toByteArray();
-        });
     }
 
 }
